Revert "internal/function: implement commit_file_stats"
This reverts commit 7ac346a32232ffefcc1d2f9f6c2b14f0a4611a5f.
