+++ internal/commitstats/commit_stats.go
@@ -7,7 +7,7 @@ import (
 	"io"
 
 	"github.com/hhatto/gocloc"
-	"github.com/src-d/enry/v2"
+	"gopkg.in/src-d/enry.v1"
 	"gopkg.in/src-d/go-git.v4"
 	"gopkg.in/src-d/go-git.v4/plumbing/object"
 	"gopkg.in/src-d/go-git.v4/utils/binary"
