+++ internal/function/commit_stats.go
@@ -8,13 +8,11 @@ import (
 	"github.com/src-d/gitbase/internal/commitstats"
 
 	"github.com/src-d/go-mysql-server/sql"
-	"gopkg.in/src-d/go-git.v4"
 	"gopkg.in/src-d/go-git.v4/plumbing"
 	"gopkg.in/src-d/go-git.v4/plumbing/object"
 )
 
-// CommitStats calculates the diff stats for a given commit. Vendored files
-// are completely ignored for the output of this function.
+// CommitStats calculates the diff stats for a given commit.
 type CommitStats struct {
 	Repository sql.Expression
 	From       sql.Expression
@@ -96,27 +94,51 @@ func (f *CommitStats) Resolved() bool {
 
 // Eval implements the Expression interface.
 func (f *CommitStats) Eval(ctx *sql.Context, row sql.Row) (interface{}, error) {
-	return evalStatsFunc(
-		ctx,
-		"commit_stats",
-		row,
-		f.Repository, f.From, f.To,
-		func(r *git.Repository, from, to *object.Commit) (interface{}, error) {
-			return commitstats.Calculate(r, from, to)
-		},
-	)
+	span, ctx := ctx.Span("gitbase.CommitStats")
+	defer span.Finish()
+
+	r, err := f.resolveRepo(ctx, row)
+	if err != nil {
+		ctx.Warn(0, "commit_stats: unable to resolve repository")
+		logrus.WithField("err", err).Error("commit_stats: unable to resolve repository")
+		return nil, nil
+	}
+
+	log := logrus.WithField("repository", r)
+
+	to, err := f.resolveCommit(ctx, r, row, f.To)
+	if err != nil {
+		ctx.Warn(0, "commit_stats: unable to resolve 'to' commit of repository: %v", r)
+		log.WithField("err", err).Error("commit_stats: unable to resolve 'to' commit")
+		return nil, nil
+	}
+
+	from, err := f.resolveCommit(ctx, r, row, f.From)
+	if err != nil {
+		ctx.Warn(0, "commit_stats: unable to resolve 'from' commit of repository: %v", r)
+		log.WithField("err", err).Error("commit_stats: unable to resolve from commit")
+		return nil, nil
+	}
+
+	result, err := commitstats.Calculate(r.Repository, from, to)
+	if err != nil {
+		ctx.Warn(0, "commit_stats: unable to calculate for repository: %v, from: %v, to: %v", r, from, to)
+		log.WithFields(logrus.Fields{
+			"err":  err,
+			"from": from,
+			"to":   to,
+		}).Error("commit_stats: unable to calculate")
+		return nil, nil
+	}
+
+	return result, nil
 }
 
-func resolveRepo(
-	ctx *sql.Context,
-	r sql.Row,
-	repo sql.Expression,
-) (*gitbase.Repository, error) {
-	repoID, err := exprToString(ctx, repo, r)
+func (f *CommitStats) resolveRepo(ctx *sql.Context, r sql.Row) (*gitbase.Repository, error) {
+	repoID, err := exprToString(ctx, f.Repository, r)
 	if err != nil {
 		return nil, err
 	}
-
 	s, ok := ctx.Session.(*gitbase.Session)
 	if !ok {
 		return nil, gitbase.ErrInvalidGitbaseSession.New(ctx.Session)
@@ -124,11 +146,8 @@ func resolveRepo(
 	return s.Pool.GetRepo(repoID)
 }
 
-func resolveCommit(
-	ctx *sql.Context,
-	r *gitbase.Repository,
-	row sql.Row,
-	e sql.Expression,
+func (f *CommitStats) resolveCommit(
+	ctx *sql.Context, r *gitbase.Repository, row sql.Row, e sql.Expression,
 ) (*object.Commit, error) {
 	str, err := exprToString(ctx, e, row)
 	if err != nil {
@@ -146,51 +165,4 @@ func resolveCommit(
 	}
 
 	return r.CommitObject(*commitHash)
-}
-
-func evalStatsFunc(
-	ctx *sql.Context,
-	name string,
-	row sql.Row,
-	repoExpr, fromExpr, toExpr sql.Expression,
-	fn func(r *git.Repository, from, to *object.Commit) (interface{}, error),
-) (interface{}, error) {
-	span, ctx := ctx.Span("gitbase." + name)
-	defer span.Finish()
-
-	r, err := resolveRepo(ctx, row, repoExpr)
-	if err != nil {
-		ctx.Warn(0, name+": unable to resolve repository")
-		logrus.WithField("err", err).Error(name + ": unable to resolve repository")
-		return nil, nil
-	}
-
-	log := logrus.WithField("repository", r)
-
-	to, err := resolveCommit(ctx, r, row, toExpr)
-	if err != nil {
-		ctx.Warn(0, name+": unable to resolve 'to' commit of repository: %v", r)
-		log.WithField("err", err).Error(name + ": unable to resolve 'to' commit")
-		return nil, nil
-	}
-
-	from, err := resolveCommit(ctx, r, row, fromExpr)
-	if err != nil {
-		ctx.Warn(0, name+": unable to resolve 'from' commit of repository: %v", r)
-		log.WithField("err", err).Error(name + ": unable to resolve from commit")
-		return nil, nil
-	}
-
-	result, err := fn(r.Repository, from, to)
-	if err != nil {
-		ctx.Warn(0, name+": unable to calculate for repository: %v, from: %v, to: %v", r, from, to)
-		log.WithFields(logrus.Fields{
-			"err":  err,
-			"from": from,
-			"to":   to,
-		}).Error(name + ": unable to calculate")
-		return nil, nil
-	}
-
-	return result, nil
 }
