+++ internal/function/is_vendor.go
@@ -3,7 +3,7 @@ package function
 import (
 	"fmt"
 
-	enry "github.com/src-d/enry/v2"
+	enry "gopkg.in/src-d/enry.v1"
 	"github.com/src-d/go-mysql-server/sql"
 	"github.com/src-d/go-mysql-server/sql/expression"
 )
