+++ internal/function/language.go
@@ -8,7 +8,7 @@ import (
 	"strconv"
 
 	lru "github.com/hashicorp/golang-lru"
-	enry "github.com/src-d/enry/v2"
+	enry "gopkg.in/src-d/enry.v1"
 	"github.com/src-d/go-mysql-server/sql"
 )
